pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["skip"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "~4.0"
black = "^22.10.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.149.2"}
strawberry-graphql-django = "^0.8.2"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.1"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"
sentry-sdk = "^1.12.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
